// backend/src/infrastructure/mailer/Model/process.ts
export enum EmailProcess {
    VERIFY_USER_EMAIL = 'VERIFY_USER_EMAIL',
    SEND_PASSWORD_RECOVERY = 'SEND_PASSWORD_RECOVERY',
    CHANGE_PASSWORD_NOTIFICATION = 'CHANGE_PASSWORD_NOTIFICATION',
    SEND_LOCK_NOTIFICATION = 'SEND_LOCK_NOTIFICATION',
    CREDIT_APPROVAL_NOTIFICATION = 'CREDIT_APPROVAL_NOTIFICATION',
    SESSION_ACTIVITY_NOTIFICATION = 'SESSION_ACTIVITY_NOTIFICATION',
    SUSPICIOUS_LOGIN_ALERT = 'SUSPICIOUS_LOGIN_ALERT',
    USER_RECOVERY = 'USER_RECOVERY',
    WELCOME_NOTIFICATION = 'WELCOME_NOTIFICATION',
    PROFILE_UPDATE_NOTIFICATION = 'PROFILE_UPDATE_NOTIFICATION', // Nuevo proceso
    GOODBYE_NOTIFICATION = 'GOODBYE_NOTIFICATION', // Nuevo proceso
  }